Added idType option to lookup method
IIRC some time ago Amazon switched to using unique identifiers for all ASINs (previously for books they used the book ISBN). We're using the asin gem to search specifically for books and therefore wanted the ability to use the lookup method in conjunction with an ISBN, and therefore I've added the idType option.